+++ composables/index.ts
@@ -1,0 +1,1 @@
+export { useSorcererBasicSkills } from './skills/sorcerer/sorcerer-basic'
